rename studio folders to dashboard
