+++ tests/helpers/dashboard/test_welcome_page.py
@@ -1,6 +1,6 @@
 import marimo as mo
 
-from dlt.helpers.studio.dlt_app import home
+from dlt.helpers.dashboard.dlt_app import home
 
 
 def test_welcome_cell():
